requirements.txt
flask
python-docx
requests
matplotlib
aiohttp